Finished Check example
This was an opportunity to introduce the Fr/ and US/ directories where
localization templates are available (for specific formats - addresses,
dates, etc…)
Also the steps in the check process includes a $ action = “” parameter
that introduces the idea of calling specific code. This is in reference
to the Smart Contract Description Language layer in CommonAccord white
paper explaining how to link the legal world to Smart Contracts.